refactor: state signature collector proxy wiring
