+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/wiring/PlatformSchedulersConfig.java
@@ -52,8 +52,7 @@ import java.time.Duration;
  * @param pcesWriter                               configuration for the preconsensus event writer scheduler
  * @param pcesSequencer                            configuration for the preconsensus event sequencer scheduler
  * @param applicationTransactionPrehandler         configuration for the application transaction prehandler scheduler
- * @param stateSignatureCollectorSchedulerType     the state signature collector scheduler type
- * @param stateSignatureCollectorUnhandledCapacity number of unhandled tasks allowed for the state signature collector
+ * @param stateSignatureCollector                  configuration for the state signature collector scheduler
  * @param shadowgraphSchedulerType                 the shadowgraph scheduler type
  * @param shadowgraphUnhandledCapacity             number of unhandled tasks allowed for the shadowgraph
  * @param consensusRoundHandlerSchedulerType       the consensus round handler scheduler type
@@ -109,8 +108,8 @@ public record PlatformSchedulersConfig(
         @ConfigProperty(defaultValue = "DIRECT") TaskSchedulerConfiguration pcesSequencer,
         @ConfigProperty(defaultValue = "CONCURRENT CAPACITY(500) FLUSHABLE UNHANDLED_TASK_METRIC")
                 TaskSchedulerConfiguration applicationTransactionPrehandler,
-        @ConfigProperty(defaultValue = "SEQUENTIAL") TaskSchedulerType stateSignatureCollectorSchedulerType,
-        @ConfigProperty(defaultValue = "500") int stateSignatureCollectorUnhandledCapacity,
+        @ConfigProperty(defaultValue = "SEQUENTIAL CAPACITY(500) FLUSHABLE UNHANDLED_TASK_METRIC")
+                TaskSchedulerConfiguration stateSignatureCollector,
         @ConfigProperty(defaultValue = "SEQUENTIAL") TaskSchedulerType shadowgraphSchedulerType,
         @ConfigProperty(defaultValue = "500") int shadowgraphUnhandledCapacity,
         @ConfigProperty(defaultValue = "SEQUENTIAL_THREAD") TaskSchedulerType consensusRoundHandlerSchedulerType,
