+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/wiring/PlatformWiring.java
@@ -22,6 +22,7 @@ import static com.swirlds.common.wiring.wires.SolderType.INJECT;
 import static com.swirlds.common.wiring.wires.SolderType.OFFER;
 import static com.swirlds.logging.legacy.LogMarker.STARTUP;
 
+import com.hedera.hapi.platform.event.StateSignaturePayload;
 import com.swirlds.base.state.Startable;
 import com.swirlds.base.state.Stoppable;
 import com.swirlds.common.context.PlatformContext;
@@ -47,6 +48,8 @@ import com.swirlds.platform.components.SavedStateController;
 import com.swirlds.platform.components.appcomm.CompleteStateNotificationWithCleanup;
 import com.swirlds.platform.components.appcomm.LatestCompleteStateNotifier;
 import com.swirlds.platform.components.consensus.ConsensusEngine;
+import com.swirlds.platform.components.transaction.system.ScopedSystemTransaction;
+import com.swirlds.platform.components.transaction.system.SystemTransactionExtractionUtils;
 import com.swirlds.platform.consensus.ConsensusSnapshot;
 import com.swirlds.platform.consensus.EventWindow;
 import com.swirlds.platform.event.AncientMode;
@@ -106,7 +109,6 @@ import com.swirlds.platform.wiring.components.RunningEventHashOverrideWiring;
 import com.swirlds.platform.wiring.components.ShadowgraphWiring;
 import com.swirlds.platform.wiring.components.StateAndRound;
 import com.swirlds.platform.wiring.components.StateHasherWiring;
-import com.swirlds.platform.wiring.components.StateSignatureCollectorWiring;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import edu.umd.cs.findbugs.annotations.Nullable;
 import java.time.Duration;
@@ -146,7 +148,7 @@ public class PlatformWiring implements Startable, Stoppable, Clearable {
     private final ComponentWiring<RoundDurabilityBuffer, List<ConsensusRound>> roundDurabilityBufferWiring;
     private final ComponentWiring<PcesSequencer, GossipEvent> pcesSequencerWiring;
     private final ComponentWiring<TransactionPrehandler, Void> applicationTransactionPrehandlerWiring;
-    private final StateSignatureCollectorWiring stateSignatureCollectorWiring;
+    private final ComponentWiring<StateSignatureCollector, List<ReservedSignedState>> stateSignatureCollectorWiring;
     private final ShadowgraphWiring shadowgraphWiring;
     private final GossipWiring gossipWiring;
     private final ComponentWiring<EventWindowManager, EventWindow> eventWindowManagerWiring;
@@ -252,7 +254,7 @@ public class PlatformWiring implements Startable, Stoppable, Clearable {
         applicationTransactionPrehandlerWiring =
                 new ComponentWiring<>(model, TransactionPrehandler.class, config.applicationTransactionPrehandler());
         stateSignatureCollectorWiring =
-                StateSignatureCollectorWiring.create(model, schedulers.stateSignatureCollectorScheduler());
+                new ComponentWiring<>(model, StateSignatureCollector.class, config.stateSignatureCollector());
         signedStateFileManagerWiring =
                 SignedStateFileManagerWiring.create(model, schedulers.signedStateFileManagerScheduler());
         stateSignerWiring = StateSignerWiring.create(schedulers.stateSignerScheduler());
@@ -465,11 +467,42 @@ public class PlatformWiring implements Startable, Stoppable, Clearable {
                 .solderTo(internalEventValidatorWiring.getInputWire(InternalEventValidator::validateEvent), INJECT);
         splitOrphanBufferOutput.solderTo(applicationTransactionPrehandlerWiring.getInputWire(
                 TransactionPrehandler::prehandleApplicationTransactions));
-        splitOrphanBufferOutput.solderTo(stateSignatureCollectorWiring.preConsensusEventInput());
-        stateSignatureCollectorWiring.getAllStatesOutput().solderTo(signedStateFileManagerWiring.saveToDiskFilter());
-        stateSignatureCollectorWiring
-                .getCompleteStatesOutput()
-                .solderTo(latestCompleteStateNexusWiring.getInputWire(LatestCompleteStateNexus::setStateIfNewer));
+
+        // From the orphan buffer, extract signatures from preconsensus events for input to the StateSignatureCollector.
+        final WireTransformer<GossipEvent, List<ScopedSystemTransaction<StateSignaturePayload>>>
+                preConsensusTransformer = new WireTransformer<>(
+                        model,
+                        "extractPreconsensusSignatureTransactions",
+                        "preconsensus signatures",
+                        event -> SystemTransactionExtractionUtils.extractFromEvent(event, StateSignaturePayload.class));
+        splitOrphanBufferOutput.solderTo(preConsensusTransformer.getInputWire());
+        preConsensusTransformer
+                .getOutputWire()
+                .solderTo(stateSignatureCollectorWiring.getInputWire(
+                        StateSignatureCollector::handlePreconsensusSignatures));
+
+        // Split output of StateSignatureCollector into single ReservedSignedStates.
+        OutputWire<ReservedSignedState> splitReservedSignedStateWire = stateSignatureCollectorWiring.getSplitOutput();
+        // Add another reservation to the signed states since we are soldering to two different input wires
+        OutputWire<ReservedSignedState> allReservedSignedStatesWire =
+                splitReservedSignedStateWire.buildAdvancedTransformer(new SignedStateReserver("allStatesReserver"));
+        // All reserved signed states are saved to disk.
+        allReservedSignedStatesWire.solderTo(signedStateFileManagerWiring.saveToDiskFilter());
+        // Filter to complete states only and add a 3rd reservation since completes states are used in two input wires.
+        OutputWire<ReservedSignedState> completeReservedSignedStatesWire = allReservedSignedStatesWire
+                .buildFilter("completeStateFilter", "states", rs -> {
+                    if (rs.get().isComplete()) {
+                        return true;
+                    } else {
+                        // close the second reservation on states that are not passed on.
+                        rs.close();
+                        return false;
+                    }
+                })
+                .buildAdvancedTransformer(new SignedStateReserver("completeStatesReserver"));
+        ;
+        completeReservedSignedStatesWire.solderTo(
+                latestCompleteStateNexusWiring.getInputWire(LatestCompleteStateNexus::setStateIfNewer));
 
         solderEventWindow();
 
@@ -533,9 +566,25 @@ public class PlatformWiring implements Startable, Stoppable, Clearable {
         signedStateHasherWiring.stateOutput().solderTo(stateSignerWiring.signState());
         signedStateHasherWiring.stateAndRoundOutput().solderTo(issDetectorWiring.stateAndRoundInput());
 
-        // FUTURE WORK: combine these two methods into a single input method, which accepts a StateAndRound object
-        signedStateHasherWiring.stateOutput().solderTo(stateSignatureCollectorWiring.getReservedStateInput());
-        signedStateHasherWiring.roundOutput().solderTo(stateSignatureCollectorWiring.getConsensusRoundInput());
+        // FUTURE WORK: combine the signedStateHasherWiring State and Round outputs into a single StateAndRound output.
+        // FUTURE WORK: Split the single StateAndRound output into separate State and Round wires.
+
+        // Extract signatures from post-consensus events for input to the StateSignatureCollector.
+        final WireTransformer<ConsensusRound, List<ScopedSystemTransaction<StateSignaturePayload>>>
+                postConsensusTransformer = new WireTransformer<>(
+                        model,
+                        "extractConsensusSignatureTransactions",
+                        "consensus events",
+                        round -> SystemTransactionExtractionUtils.extractFromRound(round, StateSignaturePayload.class));
+        signedStateHasherWiring.roundOutput().solderTo(postConsensusTransformer.getInputWire());
+        postConsensusTransformer
+                .getOutputWire()
+                .solderTo(stateSignatureCollectorWiring.getInputWire(
+                        StateSignatureCollector::handlePostconsensusSignatures));
+        // Solder the state output as input to the state signature collector.
+        signedStateHasherWiring
+                .stateOutput()
+                .solderTo(stateSignatureCollectorWiring.getInputWire(StateSignatureCollector::addReservedState));
 
         pcesWriterWiring
                 .getOutputWire()
@@ -570,10 +619,8 @@ public class PlatformWiring implements Startable, Stoppable, Clearable {
                 .statusActionOutput()
                 .solderTo(statusStateMachineWiring.getInputWire(StatusStateMachine::submitStatusAction));
 
-        stateSignatureCollectorWiring
-                .getCompleteStatesOutput()
-                .solderTo(latestCompleteStateNotifierWiring.getInputWire(
-                        LatestCompleteStateNotifier::latestCompleteStateHandler));
+        completeReservedSignedStatesWire.solderTo(latestCompleteStateNotifierWiring.getInputWire(
+                LatestCompleteStateNotifier::latestCompleteStateHandler));
 
         statusStateMachineWiring
                 .getOutputWire()
@@ -613,6 +660,7 @@ public class PlatformWiring implements Startable, Stoppable, Clearable {
         inOrderLinkerWiring.getInputWire(InOrderLinker::clear);
         roundDurabilityBufferWiring.getInputWire(RoundDurabilityBuffer::clear);
         pcesWriterWiring.getInputWire(PcesWriter::registerDiscontinuity);
+        stateSignatureCollectorWiring.getInputWire(StateSignatureCollector::clear);
     }
 
     /**
@@ -757,7 +805,7 @@ public class PlatformWiring implements Startable, Stoppable, Clearable {
      */
     @NonNull
     public InputWire<ReservedSignedState> getSignatureCollectorStateInput() {
-        return stateSignatureCollectorWiring.getReservedStateInput();
+        return stateSignatureCollectorWiring.getInputWire(StateSignatureCollector::addReservedState);
     }
 
     /**
